requirements.txt
streamlit
python-dotenv
pillow
pytesseract
requests
musicbrainzngs
spotipy
openai
